# demo/doc/demo-api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"API Programmazione film","description":"API per la visualizzazione della programmazione dei film"}],"paths":{"/demo":{"get":{"tags":["API Programmazione film"],"summary":"Descrizione dell'applicazione","description":"Gestione programmazione film","operationId":"getDescrizione","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/demo/storico":{"get":{"tags":["API Programmazione film"],"summary":"Storico programmazione","description":"Produce tutta la programmazione disponibile","operationId":"getStorico","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Programmazione"}}}}}}}},"/demo/storico/{id}":{"get":{"tags":["API Programmazione film"],"summary":"Storico specifico","description":"Produce la programmazione relativa a uno specifico identificativo","operationId":"getProgrammazioneById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Programmazione"}}}}}}},"/demo/programmazione":{"get":{"tags":["API Programmazione film"],"summary":"Programmazione settimanale","description":"Produce la programmazione della settimana in corso da lunedì a domenica","operationId":"getProgrammazione","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Programmazione"}}}}}}}},"/demo/programmazione/{dataInizio}/{dataFine}":{"get":{"tags":["API Programmazione film"],"summary":"Programmazione per periodo","description":"Produce la programmazione reperibile tra due date","operationId":"getProgrammazione_1","parameters":[{"name":"dataInizio","in":"path","required":true,"schema":{"type":"string","format":"date-time"}},{"name":"dataFine","in":"path","required":true,"schema":{"type":"string","format":"date-time"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Programmazione"}}}}}}}}},"components":{"schemas":{"Film":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"titolo":{"type":"string"},"descrizione":{"type":"string"},"regista":{"type":"string"},"anno":{"type":"integer","format":"int32"}}},"Programmazione":{"required":["data"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"data":{"type":"string","format":"date-time"},"sala":{"type":"string"},"film":{"$ref":"#/components/schemas/Film"}}}}}}